Moved mongoose stuff to tooling.ts
